// src/sections/payments_dashboard/types.ts
export interface Response {
  IsSuccess: boolean;
  StatusCode: string;
  ErrorMessages: [];
}
// ...
export interface DashboardMahsoliPaymentsResponse extends Response {
  result: {
    dashMain: {
      cntLines: number;
      smCleanWeight: number;
      smProteinDeduction: number;
      smAmountToPay: number;
      smZakatPayToDate: number;
      smVAT: number;
      smPayToDate: number;
      smADFOperationDeducted: number;
      smADFDeducted: number;
      smREDFDeducted: number;
      smDeducted: number;
      smNetAmount: number;
    };
  };
}